feat: initialize DB tables on server start
