+++ backend/src/db/tableSetup.ts
@@ -1,0 +1,102 @@
+import { query } from '../db';
+
+export const createTablesSQL = `
+  CREATE TABLE IF NOT EXISTS users (
+    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    email VARCHAR(255) UNIQUE NOT NULL,
+    password_hash VARCHAR(255) NOT NULL,
+    username VARCHAR(100),
+    avatar_url VARCHAR(255),
+    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
+  );
+
+  CREATE TABLE IF NOT EXISTS dashboards (
+    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    name VARCHAR(100) NOT NULL,
+    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    UNIQUE (user_id, name)
+  );
+
+  CREATE TABLE IF NOT EXISTS app_settings (
+    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    dashboard_id UUID NOT NULL REFERENCES dashboards(id) ON DELETE CASCADE,
+    currency VARCHAR(3) NOT NULL DEFAULT 'BRL',
+    enable_notifications BOOLEAN DEFAULT FALSE,
+    notification_time VARCHAR(5) DEFAULT '18:00', -- HH:MM
+    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    UNIQUE (user_id, dashboard_id)
+  );
+
+  CREATE TABLE IF NOT EXISTS initial_balances (
+    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    dashboard_id UUID NOT NULL REFERENCES dashboards(id) ON DELETE CASCADE,
+    balance NUMERIC(15, 2) NOT NULL,
+    currency VARCHAR(3) NOT NULL DEFAULT 'BRL',
+    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    UNIQUE (user_id, dashboard_id)
+  );
+
+  CREATE TABLE IF NOT EXISTS daily_entries (
+    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    dashboard_id UUID NOT NULL REFERENCES dashboards(id) ON DELETE CASCADE,
+    date_key DATE NOT NULL,
+    final_balance NUMERIC(15, 2) NOT NULL,
+    tags TEXT[],
+    notes TEXT,
+    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    UNIQUE (user_id, dashboard_id, date_key)
+  );
+
+  CREATE TABLE IF NOT EXISTS goals (
+    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    dashboard_id UUID NOT NULL REFERENCES dashboards(id) ON DELETE CASCADE,
+    type VARCHAR(10) NOT NULL, -- 'daily', 'weekly', 'monthly'
+    amount NUMERIC(15, 2) NOT NULL,
+    applies_to VARCHAR(20) NOT NULL, -- 'YYYY-MM-DD', 'YYYY-WNN', 'YYYY-MM'
+    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
+    CONSTRAINT check_goal_type CHECK (type IN ('daily', 'weekly', 'monthly'))
+  );
+
+  -- Trigger function to update 'updated_at' columns
+  CREATE OR REPLACE FUNCTION trigger_set_timestamp()
+  RETURNS TRIGGER AS $$
+  BEGIN
+    NEW.updated_at = NOW();
+    RETURN NEW;
+  END;
+  $$ LANGUAGE plpgsql;
+
+  -- Apply trigger to tables that have 'updated_at'
+  DO $$
+  DECLARE
+    t_name TEXT;
+  BEGIN
+    FOR t_name IN SELECT table_name FROM information_schema.columns WHERE column_name = 'updated_at' AND table_schema = current_schema()
+    LOOP
+      EXECUTE format('DROP TRIGGER IF EXISTS set_timestamp ON %I;', t_name);
+      EXECUTE format('CREATE TRIGGER set_timestamp BEFORE UPDATE ON %I FOR EACH ROW EXECUTE FUNCTION trigger_set_timestamp();', t_name);
+    END LOOP;
+  END;
+  $$;
+`;
+
+export async function createTablesIfNotExist() {
+  try {
+    await query(createTablesSQL);
+    console.log('Tables checked/created successfully.');
+  } catch (error) {
+    console.error('Error creating tables:', error);
+    throw error;
+  }
+}
